Move tasks implementations from the index file to the separate files
