+++ src/tasks/index.ts
@@ -1,0 +1,5 @@
+export * from "./compile";
+export * from "./setup";
+export * from "./make";
+export * from "./generate-verifiers";
+export * from "./clean";
