+++ src/types/tasks/generate-verifiers.ts
@@ -1,0 +1,9 @@
+import { VerifierLanguageType } from "@solarity/zkit";
+
+export type GenerateVerifiersTaskConfig = {
+  verifiersDir?: string;
+  verifiersType?: VerifierLanguageType;
+  noCompile: boolean;
+  quiet: boolean;
+  force: boolean;
+};
