+++ src/types/tasks/make.ts
@@ -1,0 +1,6 @@
+export type MakeTaskConfig = {
+  force: boolean;
+  json: boolean;
+  c: boolean;
+  quiet: boolean;
+};
