+++ src/types/tasks/setup.ts
@@ -1,0 +1,4 @@
+export type SetupTaskConfig = {
+  force: boolean;
+  quiet: boolean;
+};
